Separate file for MOVFormat class (needed for FFmpeg)
